moved addon for selenium authentication to examples/complex.
